pyproject.toml
[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[project]
name = "tablespec"
dynamic = ["version"]
description = "Python library for working with table schemas in Universal Metadata Format (UMF)"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "Erik LaBianca", email = "erik.labianca@gmail.com"},
]
keywords = ["schema", "metadata", "umf", "data-quality", "great-expectations", "pyspark"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "gitpython>=3.1.0,<4.0.0",
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "great-expectations>=1.6.0,<2.0.0",
    "openpyxl>=3.1.0",
    "typer>=0.9.0,<1.0.0",
    "rich>=13.0.0,<14.0.0",
    "ruamel.yaml>=0.18.0,<1.0.0",
]

[project.scripts]
tablespec = "tablespec.cli:app"

[project.optional-dependencies]
spark = [
    "pyspark>=4.0.0,<5.0.0",
    "pyarrow>=15.0.0,<18.0.0",
]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
fallback-version = "0.0.0"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
    "anyio>=4.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.13.3",
    "pyright>=1.1.0",
    "pre-commit>=4.0.0",
    "hypothesis>=6.0.0",
    "sqlglot>=26.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "acceptance: acceptance criteria tests tied to user stories",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:great_expectations.*",
    "ignore::DeprecationWarning:pydantic.*",
    "ignore::PendingDeprecationWarning",
]
addopts = "--strict-markers"

[tool.coverage.run]
source = ["src/tablespec"]
branch = true
omit = [
    "src/tablespec/__init__.py",
]

[tool.coverage.report]
fail_under = 70
show_missing = true
precision = 1
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "@overload",
]

[tool.uv]
package = true
